Added a simple Score Board
Added a simple Score Board on the top left hand side of the main game board. Right now it is mainly an UI component, the corresponding C# script is added but it only has a starting score of 0. The logics needs to be updated. Once finished, mini game's in-game score board could also use this templet.